ICell.asNumber return float. TableCell formats text considering if value is text, integer ou float

+++ ProjectProb/app/src/main/java/com/sharktech/projectprob/analyse/DataAnalyse.java
@@ -43,7 +43,7 @@ public class DataAnalyse {
         return mResult.get(DATA, index);
     }
 
-    public Long getFrequency(int index){
+    public float getFrequency(int index){
         return mResult.get(FREQUENCY, index).asNumber();
     }
